requirements.txt
python-dotenv==0.18.0
anthropic==0.67.0
